LayoutTests/css3/flexbox/bug605682.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>

<style>
html, body {
  margin: 0;
  padding: 0;
}

.flexbox {
  display: flex;
  background: papayawhip;
  width: 800px;
}

.flexbox > div {
  width: 300px;
  height: 300px;
  margin: 0 auto;
  background: olive;
}

.flexbox > div > div {
  overflow-y: auto;
  height: 50px;
  visibility: hidden;
}

.elm {
  height: 10px;
}

</style>
<div id=log></div>

<p>Below there should be a olive square <em>centered</em> inside a papayawhip box.</p>
<div class="flexbox">
    <div data-offset-x="250">
        <div>
            <div id="elm"></div>
        </div>
    </div>
</div>
<script>
    document.body.offsetTop;
    document.getElementById("elm").style.height = "800px";
    checkLayout('.flexbox');
</script>
